Simple board UI and KeyListener
It seems like the rest of the application for the client will need to run from the UIWindow class. It doesn't seem like the UI can be run from the main method.

public void start(Stage primaryStage) runs and launches the window. It will probably need to be treated as the main method of the client.

For now, running UIWindow creates a simple GUI with JavaFX showing the starting layout of the board. A KeyListener has been implemented that will transmit moves to the server in the future.